Phase 1: Sovereign MCP Gateway — semantic routing, cryptographic provenance, firewall, offline kill switch, integration tested
============================================================
WHAT WAS BUILT
============================================================

cortex-core
  • Config::load() parses cortex.toml with the toml crate
  • FeatureGate::from_license() maps license features to boolean flags
  • Runtime::new() and bootstrap_subsystems() initialise all Phase‑1 crates
  • Main event loop with heartbeat counter

cortex-gateway
  • EmbeddingRouter — deterministic 128‑dim bag‑of‑words vectoriser
    with L2 normalisation
  • ToolRegistry — cosine‑similarity search over registered tools
  • IntentParser — NL → ParsedIntent { action, targets } via keyword
    matching (6 actions, 16 enterprise entity types)
  • ExecutionPlanner — PlanStep per candidate tool with ATBA timeouts
  • SemanticGateway::route_intent() — full pipeline:
    parse → embed → search → construct plan
  • Axum MCP server — POST /mcp, GET /health, POST /admin/kill,
    POST /admin/revive
  • Wire‑format interop with cortex‑provenance and cortex‑security

cortex-provenance
  • TraceCapsAccumulator::attach() — BLAKE3‑hashed, parent‑linked capsules
  • MerkleChainBuilder — deterministic SHA‑256 Merkle root from leaves
  • Signer — Ed25519 key generation and signing
  • AuditLog — append‑only, queryable event log
  • ProvenanceEngine — unified provenance handle with Arc<RwLock<>>
    for safe concurrent access
  • AATFormatter — IETF AAT JSON with all 9 mandatory fields

cortex-security
  • SemanticFirewall — 6 OWASP MCP Top‑10 regex patterns blocking
    prompt injection
  • AGTPolicyEngine — Microsoft AGT bridge with default policies
  • EnforcementMode — derive PartialEq fix
  • CABPPipeline, CryptoHITL, MCIPIntegrity, MCPShield stubs
    (compilable, to be wired in Phase 2)

cortex-guard
  • CortexGuard kill switch — global AtomicBool with /admin/kill
    and /admin/revive endpoints returning 503 when active

cortex-cli
  • cortex serve --port <PORT> initialises the runtime, registers
    a demo tool, and starts the Axum MCP server

============================================================
INTEGRATION TEST RESULTS
============================================================

  Benign query    → 200  {plan: …, capsule_id: …}
  Malicious query → 403  (semantic firewall)
  Kill switch     → 503  (CortexGuard active)
  Revive          → 200  (normal service restored)

============================================================
FILES CHANGED
============================================================

  crates/cortex-core/src/config.rs            — Config::load() + test
  crates/cortex-core/src/feature_gate.rs      — FeatureGate::from_license() + tests
  crates/cortex-core/src/runtime.rs           — Runtime::new(), main loop
  crates/cortex-core/src/lib.rs               — removed duplicate tracing init
  crates/cortex-gateway/src/lib.rs            — SemanticGateway struct, Registry → Arc<RwLock<>>
  crates/cortex-gateway/src/embedding_router.rs — bag‑of‑words vectoriser + tests
  crates/cortex-gateway/src/tool_registry.rs  — cosine similarity search
  crates/cortex-gateway/src/intent_parser.rs  — NL → ParsedIntent + tests
  crates/cortex-gateway/src/execution_planner.rs — PlanStep per tool with ATBA + test
  crates/cortex-gateway/src/mcp_server.rs     — Axum router, firewall hook, kill switch
  crates/cortex-gateway/src/semantic_gateway.rs — removed (functionality in lib.rs)
  crates/cortex-gateway/Cargo.toml             — added cortex-provenance, cortex-security deps
  crates/cortex-provenance/src/lib.rs          — ProvenanceEngine, accumulator → Arc<RwLock<>>
  crates/cortex-provenance/src/tracecaps.rs   — TraceCapsAccumulator::attach() + tests
  crates/cortex-provenance/src/merkle_chain.rs — MerkleChainBuilder + tests
  crates/cortex-provenance/src/signing.rs     — Ed25519 Signer
  crates/cortex-provenance/src/audit_log.rs   — append‑only audit log
  crates/cortex-provenance/src/aat_formatter.rs — IETF AAT JSON + test
  crates/cortex-security/src/semantic_firewall.rs — OWASP patterns + tests
  crates/cortex-security/src/agt_policy_engine.rs — fixed derive conflicts, missing types
  crates/cortex-security/Cargo.toml            — added thiserror, removed cortex-gateway cycle
  crates/cortex-cli/src/main.rs                — Serve command, Tool registration with all fields, chrono dep
  crates/cortex-cli/Cargo.toml                 — added axum, chrono deps

+++ crates/cortex-security/src/semantic_firewall.rs
@@ -1,3 +1,4 @@
+#![allow(unused_imports, dead_code, unused_variables)]
 use crate::SecurityError;
 use regex::RegexSet;
 use blake3::Hasher;
@@ -132,7 +133,7 @@ impl SemanticFirewall {
         }
 
         ToolDescriptorScanResult {
-            risk_score: risk.min(1.0),
+            risk_score: (risk as f64).min(1.0),
             threats,
             passed: risk < 0.5,
         }
